+++ modules/create_comment.js
@@ -1,10 +1,10 @@
-  const createComment = (el) => {
-    const li = document.createElement('li');
-    li.innerHTML = `<span>(${el.creation_date})</span>
+const createComment = (el) => {
+  const li = document.createElement('li');
+  li.innerHTML = `<span>(${el.creation_date})</span>
     <span>${el.username} said: </span>
     <span>${el.comment}</span>
-    `
-    return li;
-}
+    `;
+  return li;
+};
 
 export default createComment;
