+++ modules/event_listeners.js
@@ -1,9 +1,5 @@
 import LikesService from './likes_service';
 import createPopUp from './popup_ui';
-import CommentService from './comment__service.js';
-import createComment from './create_comment';
-
-let savedId = '';
 
 const popup = document.querySelector('.comment__popup');
 
@@ -30,8 +26,6 @@ const handleLike = async (event) => {
 };
 /** @param {Event} data */
 const handleComment = (data) => {
-  console.log(data.target);
-  savedId = data.target.id;
   createPopUp(data.target.id);
   popup.style.display = 'block';
 };
@@ -40,28 +34,4 @@ togglePopup.addEventListener('click', () => {
   popup.style.display = 'none';
 });
 
-
-const handleSubmit = (event) => {
-  event.preventDefault();
-  /**
-   * @type {HTMLFormElement}
-   */
-  const form = event.target;
-    const name = form.elements.name.value
-    const comment = form.elements.comment.value;
-   
-    if (name !== '' && comment !== '' && savedId !== '') {
-      const data = {
-        item_id: savedId,
-        username: name,
-        comment: comment,
-      };
-      form.previousElementSibling.appendChild(createComment({creation_date:"1s Ago",username:name,comment:comment}))
-      CommentService.postComments(data);
-      form.elements.name.value = '';
-      form.elements.comment.value = '';
-    }
-  
-};
-
-export { handleComment, handleLike, handleSubmit };
+export { handleComment, handleLike };
